Move activity.ts from lib/ -> client/
