+++ packages/shared/src/logic/index.ts
@@ -3,7 +3,6 @@ export * from '@tloncorp/api/lib/utils';
 export * from '@tloncorp/api/lib/references';
 export * from './embed';
 export * from '@tloncorp/api/lib/types';
-export * from '@tloncorp/api/lib/activity';
 export * from '@tloncorp/api/lib/branch';
 export * from '@tloncorp/api/lib/deeplinks';
 export * as featureFlags from '@tloncorp/api/lib/featureFlags';
@@ -15,3 +14,4 @@ export * from '@tloncorp/api/lib/wayfinding';
 export * from '@tloncorp/api/lib/postContent';
 export * from '@tloncorp/api/lib/content-helpers';
 export * from '@tloncorp/api/lib/pinning';
+export * from '@tloncorp/api/client/activity';
